#106.558: Bug: '0' is a legal plan value

+++ lib/class/Api/Plan/Set/cashApiPlanSetHandler.class.php
@@ -22,7 +22,7 @@ class cashApiPlanSetHandler implements cashApiHandlerInterface
 
         $model = cash()->getModel('cashPlan');
 
-        if (empty($request->amount)) {
+        if (empty($request->amount) && $request->amount != 0) {
             $model->deleteByField($data);
             return [];
         } elseif ($plan = $model->getByField($data)) {
